Fix button height and gutter
When multiple line text was provided to the Button, use of line-height was causing a bad looking situation. Changed from line-height to max-height and added padding to text div to maintain spacing between top/bottom of button and text. Also added gutter to both sides to maintain a centered text regardless of icon/no icon.